Create OOPS Assignment-1
Q1. Explain Class and Object with respect to Object-Oriented Programming. Give a suitable example.

Objects have states and behaviors. Example: A dog has states - color, name, breed as well as behaviors − wagging the tail, barking, eating. An object is an instance of a class. Class − A class can be defined as a template/blueprint that describes the behavior/state that the object of its type support.


Q2. Name the four pillars of OOPs.

The Four Pillars of Object Oriented Programming
Abstraction.
Encapsulation.
Inheritance.
Polymorphism.

Q3. Explain why the __init__() function is used. Give a suitable example.

The __init__ method is the Python equivalent of the C++ constructor in an object-oriented approach. The __init__ function is called every time an object is created from a class. The __init__ method lets the class initialize the object's attributes and serves no other purpose. It is only used within classes

Q4. Why self is used in OOPs?

Python does not use the @ syntax to refer to instance attributes.

Q5. What is inheritance? Give an example for each type of inheritance.

Inheritance is a core principle of object-oriented programming where a class can inherit properties and behavior from another class.